src/third_party/blink/web_tests/http/tests/loading/module-script-wrong-mime-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
CONSOLE ERROR: Failed to load module script: The server responded with a non-JavaScript, CSS MIME type of "text/plain". Strict MIME type checking is enforced for module scripts per HTML spec.
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
PASS module script with invalid mime type shouldn't have run.
Harness: the test ran to completion.
